% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_P_from_blockmodels.R
\name{A_P_from_blockmodels}
\alias{A_P_from_blockmodels}
\title{Extract adjacency and probability matrix from blockmodels object}
\usage{
A_P_from_blockmodels(model)
}
\arguments{
\item{model}{a blockmodels object}
}
\value{
adjacency matrix (A), probability matrix (P) and model type
("bernoulli" or "poisson")
}
\description{
Extract adjacency and probability matrix from blockmodels object
}
\examples{
library(blockmodels)
npc <- 30 # nodes per class
Q <- 3 # classes
n <- npc * Q # nodes
Z<-diag(Q)\%x\%matrix(1,npc,1)
P<-matrix(runif(Q*Q),Q,Q)
M<-1*(matrix(runif(n*n),n,n)<Z\%*\%P\%*\%t(Z)) ## adjacency matrix
fit <- BM_bernoulli("SBM",M, plotting='')
out = A_P_from_blockmodels(model=fit)
out$A
out$P
out$modeltype
}
