% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_P_from_ergm.R
\name{A_P_from_ergm}
\alias{A_P_from_ergm}
\title{Extract adjacency and probability matrix from ergm object}
\usage{
A_P_from_ergm(model)
}
\arguments{
\item{model}{an ergm object}
}
\value{
adjacency matrix (A) and probability matrix (P)
}
\description{
Extract adjacency and probability matrix from ergm object
}
\examples{
library(ergm)
data(faux.dixon.high)
fit <- ergm(faux.dixon.high ~ edges + mutual)
out = A_P_from_ergm(model=fit)
out$A
out$P
}
