% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_P_from_fastRG.R
\name{A_P_from_fastRG}
\alias{A_P_from_fastRG}
\title{Extract adjacency matrix and probability matrix from fastRG package}
\usage{
A_P_from_fastRG(model)
}
\arguments{
\item{model}{can be objects from fastRG package including: "undirected_dcsbm",
"directed_dcsbm", "undirected_sbm"}
}
\value{
adjacency matrix (A), probability matrix (P) and model type ("poisson" or "binomial")
}
\description{
Only applicable for stochastic block model and degree corrected
stochastic block model. Probability matrix (P) is obtained by maximising likelihood.
}
\examples{
set.seed(27)
k <- 5
n <- 100
B <- matrix(stats::runif(k * k), nrow = k, ncol = k)
theta <- round(stats::rlnorm(n, 2))
pi <- c(1, 2, 4, 1, 1)
m1 <- fastRG::dcsbm(theta = theta, B = B, pi = pi, expected_degree = 50)
out = A_P_from_fastRG(model = m1)
out$A
out$P
out$modeltype
}
