% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_P_from_greed.R
\name{A_P_from_greed}
\alias{A_P_from_greed}
\title{Extract adjacency and probability matrix from SbmFit and DcSbmFit object}
\usage{
A_P_from_greed(data, blockmodel)
}
\arguments{
\item{data}{network connection data with type "matrix", "dgCmatrix" or "igraph"}

\item{blockmodel}{an SbmFit or DcSbmFit object}
}
\value{
adjacency matrix (A), probability matrix (P) and model type ("bernoulli"
or "poisson")
}
\description{
Extract adjacency and probability matrix from SbmFit and DcSbmFit object
from "greed" package
}
\examples{
library(greed)
data(Books)
sbm = greed(Books$X, model = Sbm())
out1 = A_P_from_greed(data=Books$X, blockmodel=sbm)
library(igraphdata)
data(karate)
dcsbm = greed(karate, model=DcSbm())
out2 = A_P_from_greed(data=karate, blockmodel=dcsbm)
}
