% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_P_from_igraph.R
\name{A_P_from_igraph}
\alias{A_P_from_igraph}
\title{Extract adjacency and probability matrix from igraph object}
\usage{
A_P_from_igraph(graph)
}
\arguments{
\item{graph}{an igraph object with name "Stochastic block model"}
}
\value{
adjacency matrix (A) and probability matrix (P)
}
\description{
Only applicable for stochastic block model. Number of communities
is obtained from greedy optimization of modularity. Community labels are obtained
from spectral clustering algorithm.
}
\examples{
num = 1000
pm = matrix(c(.5, .1, .1, .5), 2, 2)
bs = c(300, 700)
g = igraph::sample_sbm(num, pm, bs)
out = A_P_from_igraph(graph = g)
out$A
out$P
}
