% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccr.R
\name{ccr}
\alias{ccr}
\title{Calculate Correct Classification Rate}
\usage{
ccr(label_true, label_est, parallel = F)
}
\arguments{
\item{label_true}{true community labels}

\item{label_est}{estimated community labels}

\item{parallel}{allow use parallel computing, default is FALSE}
}
\value{
Correct Classification Rate
}
\description{
Calculate Correct Classification Rate
}
\examples{
library(csgc)
label_true = sample(1:5,100,T)
label_est = sample(1:5,100,T)
ccr(label_true,label_est)
}
