% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csgc.R
\name{csgc}
\alias{csgc}
\title{Calculate basic centred subgraph counts}
\usage{
csgc(A, P, var.structure = c("binomial", "poisson"))
}
\arguments{
\item{A}{adjacency matrix}

\item{P}{probability matrix}

\item{var.structure}{allow choose of binomial/poisson model.}
}
\value{
Unnormalised statistics (V.graph), variance of statistics (vr.graph),
normalised statistics (t.graph)
}
\description{
Calculate basic centred subgraph counts
}
\examples{
library(csgc)
k = 4
n = 100
prob = c(0.8, 0.1)
K = matrix(prob[2],k,k)
diag(K) = prob[1]
z = rep(1:k,each=n/k)
mat = gen_adj_sbm(K,z)
A = mat$A
P = mat$P
csgc(A,P,"binomial")
}
