% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csgc_greedy.R
\name{csgc_greedy}
\alias{csgc_greedy}
\title{Csgc greedy algorithm}
\usage{
csgc_greedy(A, z0, var.structure = "binomial", parallel = F)
}
\arguments{
\item{A}{adjacency matrix}

\item{z0}{initial community labels}

\item{var.structure}{allow choose of binomial/poisson model, default is binomial}

\item{parallel}{allow use parallel computing, default is FALSE}
}
\value{
A list of initial/optimized/record values
\itemize{
\item zin - initial community labels.
\item chisqin - initial chi-square statistic.
\item statsin - initial csgc statistics.
\item zout - optimised community labels.
\item chisqout - optimised chi-square statistic.
\item statsout - optimised csgc statistics.
\item zlog - record of community labels at each step.
\item chisqlog - record of chi-square statistic at each step.
}
}
\description{
Use csgc greedy algorithm to obtain optimised community labels
and the corresponding chi-square statistics, under stochastic block model
}
\examples{
library(csgc)
k = 2
n = 400
K = matrix(c(0.8, 0.5, 0.5, 0.1), k, k)
z = rep(1:k,each=n/k)
A = gen_adj_sbm(K,z)$A
z0 = spectral_sbm(A,k)
csgc_greedy(A,z0)
}
