% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcsbm_mle.R
\name{dcsbm_mle}
\alias{dcsbm_mle}
\title{Calculate Maximum Likelihood Estimator of Degree Corrected Stochastic Block Model}
\usage{
dcsbm_mle(A, z)
}
\arguments{
\item{A}{adjacency matrix}

\item{z}{community labels}
}
\value{
estimated connection probability matrix (K) and estimated probability matrix (P)
}
\description{
Calculate Maximum Likelihood Estimator of Degree Corrected Stochastic
Block Model, using poisson model assuming simple graph without self-loop.
}
\examples{
library(csgc)
k = 4
n = 200
prob = c(0.8, 0.1)
K = matrix(prob[2],k,k)
diag(K) = prob[1]
Theta = runif(n,0.8,1)
z = rep(1:k,each=n/k)
A = gen_adj_dcsbm(K,Theta,z)$A
dcsbm_mle(A,z)
}
