% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_adj_dcsbm.R
\name{gen_adj_dcsbm}
\alias{gen_adj_dcsbm}
\title{Generate adjacency matrix from Degree Corrected Stochastic Block Model}
\usage{
gen_adj_dcsbm(K, Theta, z)
}
\arguments{
\item{K}{connection probability matrix}

\item{Theta}{degree parameter vector}

\item{z}{community labels}
}
\value{
adjacency matrix (A) and probability matrix (P)
}
\description{
Generate adjacency matrix from Degree Corrected Stochastic Block Model
}
\examples{
library(csgc)
k = 4
n = 200
prob = c(0.8, 0.1)
K = matrix(prob[2],k,k)
diag(K) = prob[1]
Theta = runif(n,0.8,1)
z = rep(1:k,each=n/k)
gen_adj_dcsbm(K,Theta,z)
}
