% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_mle.R
\name{sbm_mle}
\alias{sbm_mle}
\title{Calculate Maximum Likelihood Estimator of Stochastic Block Model}
\usage{
sbm_mle(A, z)
}
\arguments{
\item{A}{adjacency matrix}

\item{z}{community labels}
}
\value{
estimated connection probability matrix (K) and estimated probability matrix (P)
}
\description{
Calculate Maximum Likelihood Estimator of Stochastic Block Model,
using bernoulli model assuming simple graph without self-loop.
}
\examples{
library(csgc)
k = 4
n = 200
prob = c(0.8, 0.1)
K = matrix(prob[2],k,k)
diag(K) = prob[1]
z = rep(1:k,each=n/k)
A = gen_adj_sbm(K,z)$A
sbm_mle(A,z)
}
