% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_sbm.R
\name{spectral_sbm}
\alias{spectral_sbm}
\title{Spectral clustering for Stochastic Block Model}
\usage{
spectral_sbm(A, k)
}
\arguments{
\item{A}{adjacency matrix}

\item{k}{number of communities}
}
\value{
estimated labels
}
\description{
Tianxi Li, Elizaveta Levina, Ji Zhu and Can M. Le (2022). randnet:
Random Network Model Estimation, Selection and Parameter Tuning. R
package version 0.5. https://CRAN.R-project.org/package=randnet
}
\examples{
library(csgc)
k = 4
n = 200
prob = c(0.8, 0.1)
K = matrix(prob[2],k,k)
diag(K) = prob[1]
z = rep(1:k,each=n/k)
A = gen_adj_sbm(K,z)$A
spectral_sbm(A,k)
}
