% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{trace}
\alias{trace}
\title{Return the trace of a matrix}
\usage{
trace(A)
}
\arguments{
\item{A}{matrix}
}
\value{
trace of matrix A
}
\description{
Return the trace of a matrix
}
\examples{
library(csgc)
A = matrix(c(1,2,3,4),2,2)
trace(A)
}
