% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasserstein_uniform.R
\name{wasserstein_uniform}
\alias{wasserstein_uniform}
\title{Calculate Wasserstein distance between uniform distribution and target distribution}
\usage{
wasserstein_uniform(p)
}
\arguments{
\item{p}{a sequence of values from 0 to 1}
}
\value{
2 times the Wasserstein distance
}
\description{
Calculate Wasserstein distance between uniform distribution on (0,1) and
the target distribution from 0 to 1 (for example a sequence of p-values)
}
\examples{
set.seed(123)
library(csgc)
p = runif(100)
wasserstein_uniform(p)
}
