% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_dcsbm.R
\name{spectral_dcsbm}
\alias{spectral_dcsbm}
\title{Spectral clustering for Degree Corrected Stochastic Block Model}
\usage{
spectral_dcsbm(A, k)
}
\arguments{
\item{A}{adjacency matrix}

\item{k}{number of communities}
}
\value{
estimated labels
}
\description{
Spectral clustering for Degree Corrected Stochastic Block Model
}
\examples{
library(csgc)
k = 4
n = 200
prob = c(0.8, 0.1)
K = matrix(prob[2],k,k)
diag(K) = prob[1]
Theta = runif(n,0.8,1)
z = rep(1:k,each=n/k)
A = gen_adj_dcsbm(K,Theta,z)$A
spectral_dcsbm(A,k)
}
